situate new hosting types, remove ad hoc pixel sizings for provider logos
