+++ packages/api/src/client/hostingApi.ts
@@ -11,6 +11,35 @@ import {
   User,
   getConstants,
 } from '../types';
+import type {
+  HostingHeartBeatCode,
+  TlawnBotInfo,
+  TlawnConfig,
+  TlawnOAuthProvider,
+  TlawnOAuthStartRequest,
+  TlawnOAuthStartResponse,
+  TlawnOAuthStatus,
+  TlawnPrimaryModelUpdate,
+  TlawnProviderConfigInfo,
+  TlawnProviderModel,
+} from '../types/hosting';
+
+export type {
+  HostingHeartBeatCode,
+  TlawnBotInfo,
+  TlawnConfig,
+  TlawnModelEntry,
+  TlawnOAuthGrant,
+  TlawnOAuthProvider,
+  TlawnOAuthProviderKind,
+  TlawnOAuthStartRequest,
+  TlawnOAuthStartResponse,
+  TlawnOAuthStatus,
+  TlawnOAuthUpstream,
+  TlawnPrimaryModelUpdate,
+  TlawnProviderConfigInfo,
+  TlawnProviderModel,
+} from '../types/hosting';
 
 const logger = createDevLogger('hostingApi', false);
 interface StoredValue<T> {
@@ -225,94 +254,8 @@ function jsonInit(method: string, body: unknown): RequestInit {
   };
 }
 
-// --- Tlawn (bot) types ---
-
-export interface TlawnProviderConfigInfo {
-  keys: Record<string, string>;
-  models: TlawnModelEntry[];
-  defaultKeys: Record<string, { key: string; id?: string }>;
-}
-
-export interface TlawnModelEntry {
-  provider: string;
-  model: string;
-}
-
-export interface TlawnPrimaryModelUpdate {
-  provider: string;
-  model: string;
-  fallbacks?: TlawnModelEntry[];
-}
-
-export interface TlawnBotInfo {
-  enabled: boolean;
-  provider?: string;
-  model?: string;
-  moon?: string;
-}
-
-export interface TlawnConfig {
-  dmAllowlist: string[];
-  defaultAuthorizedShips: string[];
-  channelRules: Record<string, { mode: string; allowedShips: string[] }>;
-  groupChannels: string[];
-  groupInviteAllowlist: string[];
-  autoAcceptDmInvites: boolean;
-  autoDiscoverChannels: boolean;
-}
-
-export interface TlawnOAuthGrant {
-  connected: boolean;
-  expired: boolean;
-  expiresAt: string | null;
-  hasRefreshToken: boolean;
-  provider: string;
-  scopes: string;
-  tokenType: string;
-}
-
-export interface TlawnOAuthStatus {
-  available: boolean;
-  grants: TlawnOAuthGrant[];
-}
-
-export type TlawnOAuthProviderKind = 'standard' | 'mcp_remote';
-
-export type TlawnOAuthUpstream =
-  | {
-      mode: 'proxy';
-      name: string;
-      url: string;
-    }
-  | {
-      mode: 'openapi';
-      name: string;
-      schemaUrl: string;
-    };
-
-export interface TlawnOAuthProvider {
-  authUrl?: string;
-  displayName: string;
-  id: string;
-  kind: TlawnOAuthProviderKind;
-  revokeUrl?: string;
-  scopes: string;
-  suggestedUpstream: TlawnOAuthUpstream;
-  template: string;
-  tokenUrl?: string;
-}
-
 interface TlawnOAuthProvidersResponse {
   providers: TlawnOAuthProvider[];
-}
-
-export interface TlawnOAuthStartRequest {
-  providerId: string;
-  finalRedirectUrl: string;
-}
-
-export interface TlawnOAuthStartResponse {
-  authUrl: string;
 }
 
 interface RawTlawnOAuthStartResponse {
@@ -362,11 +305,6 @@ export async function setTlawnPrimaryModel(
     `/v1/tlawn/users/${userId}/primary-model`,
     jsonInit('PUT', update)
   );
-}
-
-export interface TlawnProviderModel {
-  id: string;
-  [key: string]: unknown;
 }
 
 /**
@@ -632,7 +570,6 @@ export async function awaitBotRunning(
   return false;
 }
 
-export type HostingHeartBeatCode = 'expired' | 'ok' | 'unknown';
 export const getHostingHeartBeat = async (): Promise<HostingHeartBeatCode> => {
   const userId = await sessionStore.userId.getValue();
   const response = await rawHostingFetch(`/v1/users/${userId}`);
