+++ packages/api/src/types/hosting.ts
@@ -66,6 +66,97 @@ export enum HostedNodeStatus {
   Unknown = 'Unknown',
 }
 
+export type HostingHeartBeatCode = 'expired' | 'ok' | 'unknown';
+
+export interface TlawnProviderConfigInfo {
+  keys: Record<string, string>;
+  models: TlawnModelEntry[];
+  defaultKeys: Record<string, { key: string; id?: string }>;
+}
+
+export interface TlawnModelEntry {
+  provider: string;
+  model: string;
+}
+
+export interface TlawnPrimaryModelUpdate {
+  provider: string;
+  model: string;
+  fallbacks?: TlawnModelEntry[];
+}
+
+export interface TlawnProviderModel {
+  id: string;
+  [key: string]: unknown;
+}
+
+export interface TlawnBotInfo {
+  enabled: boolean;
+  provider?: string;
+  model?: string;
+  moon?: string;
+}
+
+export interface TlawnConfig {
+  dmAllowlist: string[];
+  defaultAuthorizedShips: string[];
+  channelRules: Record<string, { mode: string; allowedShips: string[] }>;
+  groupChannels: string[];
+  groupInviteAllowlist: string[];
+  autoAcceptDmInvites: boolean;
+  autoDiscoverChannels: boolean;
+}
+
+export interface TlawnOAuthGrant {
+  connected: boolean;
+  expired: boolean;
+  expiresAt: string | null;
+  hasRefreshToken: boolean;
+  provider: string;
+  scopes: string;
+  tokenType: string;
+}
+
+export interface TlawnOAuthStatus {
+  available: boolean;
+  grants: TlawnOAuthGrant[];
+}
+
+export type TlawnOAuthProviderKind = 'standard' | 'mcp_remote';
+
+export type TlawnOAuthUpstream =
+  | {
+      mode: 'proxy';
+      name: string;
+      url: string;
+    }
+  | {
+      mode: 'openapi';
+      name: string;
+      schemaUrl: string;
+    };
+
+export interface TlawnOAuthProvider {
+  authUrl?: string;
+  displayName: string;
+  id: string;
+  kind: TlawnOAuthProviderKind;
+  revokeUrl?: string;
+  scopes: string;
+  suggestedUpstream: TlawnOAuthUpstream;
+  template: string;
+  tokenUrl?: string;
+}
+
+export interface TlawnOAuthStartRequest {
+  providerId: string;
+  finalRedirectUrl: string;
+}
+
+export interface TlawnOAuthStartResponse {
+  authUrl: string;
+}
+
 export function nodeUrlIsHosted(url: string) {
   return url.endsWith('tlon.network') || url.endsWith('.test.tlon.systems');
 }
